requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements/production.txt --strip-extras requirements/production.in
#
aiofiles==23.2.1
    # via -r requirements/base.txt
aiosqlite==0.17.0
    # via
    #   -r requirements/base.txt
    #   tortoise-orm
asyncpg==0.28.0
    # via -r requirements/base.txt
beautifulsoup4==4.12.2
    # via -r requirements/base.txt
certifi==2023.7.22
    # via
    #   -r requirements/base.txt
    #   sentry-sdk
dnspython==2.0.0
    # via
    #   -r requirements/base.txt
    #   ipwhois
graphviz==0.20.1
    # via -r requirements/base.txt
humanize==4.8.0
    # via -r requirements/base.txt
idna==3.4
    # via
    #   -r requirements/base.txt
    #   yarl
ipwhois==1.2.0
    # via -r requirements/base.txt
iso8601==1.1.0
    # via
    #   -r requirements/base.txt
    #   tortoise-orm
jaraco-context==4.3.0
    # via
    #   -r requirements/base.txt
    #   wolframalpha
markdown-it-py==3.0.0
    # via
    #   -r requirements/base.txt
    #   rich
mdurl==0.1.2
    # via
    #   -r requirements/base.txt
    #   markdown-it-py
more-itertools==10.1.0
    # via
    #   -r requirements/base.txt
    #   wolframalpha
mpmath==1.3.0
    # via
    #   -r requirements/base.txt
    #   sympy
multidict==6.0.4
    # via
    #   -r requirements/base.txt
    #   yarl
pillow==10.0.0
    # via -r requirements/base.txt
psutil==5.9.5
    # via -r requirements/base.txt
pygments==2.16.1
    # via
    #   -r requirements/base.txt
    #   rich
pypika-tortoise==0.1.6
    # via
    #   -r requirements/base.txt
    #   tortoise-orm
python-json-logger==2.0.7
    # via -r requirements/base.txt
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   tortoise-orm
redis==5.0.0
    # via -r requirements/base.txt
rich==13.5.2
    # via -r requirements/base.txt
sentry-sdk==1.29.2
    # via -r requirements/base.txt
soupsieve==2.4.1
    # via
    #   -r requirements/base.txt
    #   beautifulsoup4
statsd==4.0.1
    # via -r requirements/base.txt
sympy==1.12
    # via -r requirements/base.txt
tortoise-orm==0.20.0
    # via -r requirements/base.txt
typing-extensions==4.7.1
    # via
    #   -r requirements/base.txt
    #   aiosqlite
urllib3==2.0.4
    # via
    #   -r requirements/base.txt
    #   sentry-sdk
wolframalpha==5.0.0
    # via -r requirements/base.txt
xmltodict==0.13.0
    # via
    #   -r requirements/base.txt
    #   wolframalpha
yarl==1.9.2
    # via -r requirements/base.txt
